Amend lookup of the reference rate
The BoE "Reference rate" is defined as the rate in effect on the last day of the preceding 6-month period.

The periods are 1st Jan - 30th June  and 1st July - 31st December (inclusive), therefore, a _Due Date_ of July 15th 2023 would use the base rate in effect on 30th June 2023.

Closes #127

This patch fetches the correct reference rate in effect at the time the debt became due/late and updates the response to include the reference date so that it is clear which date is being used to determine the rate in effect.